Expose SerializeExtraArgs functions via ccipevm/ccipsolana packages
