+++ core/capabilities/ccip/ccipevm/gas_helpers_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/latest/message_hasher"
+	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/v1_6_0/message_hasher"
 	"github.com/smartcontractkit/chainlink-ccip/pkg/types/ccipocr3"
 )
 
@@ -166,19 +166,6 @@ func makeExtraArgsV2(gasLimit uint64, allowOOO bool) []byte {
 		panic(err)
 	}
 	return extraArgs
-}
-
-func SerializeExtraArgs(tag []byte, method string, inputs ...any) ([]byte, error) {
-	v, err := messageHasherABI.Methods[method].Inputs.Pack(inputs...)
-	return append(tag, v...), err
-}
-
-func SerializeEVMExtraArgsV1(data message_hasher.ClientEVMExtraArgsV1) ([]byte, error) {
-	return SerializeExtraArgs(evmExtraArgsV1Tag, "encodeEVMExtraArgsV1", data)
-}
-
-func SerializeClientGenericExtraArgsV2(data message_hasher.ClientGenericExtraArgsV2) ([]byte, error) {
-	return SerializeExtraArgs(evmExtraArgsV2Tag, "encodeGenericExtraArgsV2", data)
 }
 
 func getTokenAmounts(t *testing.T, numTokens int, tokenGasOverhead uint32) []ccipocr3.RampTokenAmount {
